+++ source/selects/bountyedit.js
@@ -1,7 +1,7 @@
 const { ModalBuilder, TextInputBuilder, ActionRowBuilder, TextInputStyle, GuildScheduledEventEntityType } = require('discord.js');
 const { SelectWrapper } = require('../classes');
 const { YEAR_IN_MS } = require('../constants');
-const { timeConversion, checkTextsInAutoMod } = require('../util/textUtil');
+const { timeConversion, checkTextsInAutoMod, trimForModalTitle } = require('../util/textUtil');
 
 const mainId = "bountyedit";
 module.exports = new SelectWrapper(mainId, 3000,
@@ -33,7 +33,7 @@ module.exports = new SelectWrapper(mainId, 3000,
 		}
 		interaction.showModal(
 			new ModalBuilder().setCustomId(interaction.id)
-				.setTitle(`Editing Bounty (${bounty.title})`)
+				.setTitle(trimForModalTitle(`Edit Bounty: ${bounty.title}`))
 				.addComponents(
 					new ActionRowBuilder().addComponents(
 						new TextInputBuilder().setCustomId("title")